Rename zipline-security to zipline-cryptography
