+++ settings.gradle.kts
@@ -4,13 +4,13 @@ include(":zipline")
 include(":zipline-api-validator")
 include(":zipline-bytecode")
 include(":zipline-cli")
+include(":zipline-cryptography")
 include(":zipline-gradle-plugin")
 include(":zipline-kotlin-plugin")
 include(":zipline-kotlin-plugin-tests")
 include(":zipline-loader")
 include(":zipline-loader-testing")
 include(":zipline-profiler")
-include(":zipline-security")
 include(":zipline-testing")
 
 enableFeaturePreview("TYPESAFE_PROJECT_ACCESSORS")
