+++ zipline-cryptography/api/zipline-cryptography.api
@@ -1,0 +1,4 @@
+public final class app/cash/zipline/cryptography/ZiplineSecurityJvmKt {
+	public static final fun installCryptographyService (Lapp/cash/zipline/Zipline;)V
+}
+
