+++ zipline-cryptography/src/jvmMain/kotlin/app/cash/zipline/cryptography/ZiplineSecurityJvm.kt
@@ -1,0 +1,23 @@
+/*
+ * Copyright (C) 2024 Cash App
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package app.cash.zipline.cryptography
+
+import app.cash.zipline.Zipline
+import java.security.SecureRandom
+
+actual fun Zipline.installCryptographyService() {
+  installCryptographyServiceInternal(RealZiplineCryptographyService(SecureRandom()))
+}
