--- zipline-security/README.md
@@ -1,3 +1,0 @@
-# Zipline Security
-
-This implements cryptography APIs for use by guest code.
