--- zipline-security/api/zipline-security.api
@@ -1,4 +1,0 @@
-public final class app/cash/zipline/security/ZiplineSecurityJvmKt {
-	public static final fun installSecurityService (Lapp/cash/zipline/Zipline;)V
-}
-
