--- zipline-security/src/jvmMain/kotlin/app/cash/zipline/security/RealZiplineSecurityService.kt
@@ -1,28 +1,0 @@
-/*
- * Copyright (C) 2024 Cash App
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package app.cash.zipline.security
-
-import java.security.SecureRandom
-
-internal class RealZiplineSecurityService(
-  private val secureRandom: SecureRandom,
-) : ZiplineSecurityService {
-  override fun nextSecureRandomBytes(size: Int): ByteArray {
-    val result = ByteArray(size)
-    secureRandom.nextBytes(result)
-    return result
-  }
-}
